+++ src/lib.rs
@@ -1,15 +1,19 @@
+// #![deny(missing_docs)]
+#![deny(missing_debug_implementations)]
+#![cfg_attr(test, deny(warnings))]
+#![cfg_attr(docs_rs_workaround, feature(extern_prelude))]
 //! # sparkpost
 //!
 //! Rust bindings for sparkpost email api v1
 //! ## WARNING!
 //! ### Work in Progress, Expect breaking changes
 
-extern crate reqwest;
-extern crate serde_json;
-
 #[macro_use]
 extern crate serde_derive;
+extern crate chrono;
+extern crate reqwest;
 extern crate serde;
+extern crate serde_json;
 
 #[cfg(test)]
 extern crate base64;